% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendClimateAssessmentToMif.R
\name{appendClimateAssessmentToMif}
\alias{appendClimateAssessmentToMif}
\title{Append Climate Assessment Data to MIF File}
\usage{
appendClimateAssessmentToMif(qf, mif)
}
\arguments{
\item{qf}{`quitte` data frame containing the climate assessment data to be appended}

\item{mif}{File path to the MIF to which the data will be appended}
}
\value{
The function returns the result of writing the updated data to the MIF file.
}
\description{
This function appends climate assessment data to a MIF (Model Intercomparison Framework) file.
It removes data from old MAGICC7 runs to avoid duplication and then writes the updated data back to the MIF file.
}
