% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archiveClimateAssessmentData.R
\name{archiveClimateAssessmentData}
\alias{archiveClimateAssessmentData}
\title{Archive Climate Assessment Data}
\usage{
archiveClimateAssessmentData(
  climateDir,
  climateArchiveDir = NULL,
  extraFiles = c(),
  returnFn = TRUE
)
}
\arguments{
\item{climateDir}{Base directory containing climate assessment data}

\item{climateArchiveDir}{Directory where the climate assessment data will be archived. Default is NULL, in which case
a new 'archive' subdir is created in the climate directory. Note: This directory is NOT deleted after archiving,
only the temporary directory created within it is}

\item{extraFiles}{Character vector of additional file paths to include in the archive. Default is an empty vector}

\item{returnFn}{A logical value indicating whether to return the path to the tar file. Defaults to TRUE}
}
\value{
If returnFn is TRUE, the function returns the path to the tar file. Otherwise, it returns invisible()
}
\description{
Archives climate assessment data by copying all *.csv/*.xlsx from climate diretory (and, if specified, extra files)
to a temporary directory, then tar'ing and gzip'ing the temporary directory & deleting it afterwards
}
\examples{
\dontrun{
cfg <- climateAssessmentConfig("<outputDir>", "iteration")
extraFiles <- c("pm_globalMeanTemperature", "p15_forc_magicc")
archiveClimateAssessmentData(cfg$climateArchiveDir, extraFiles)
}
}
