% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpToGdx.R
\name{dumpToGdx}
\alias{dumpToGdx}
\title{assessmentDataToGDX}
\usage{
dumpToGdx(relevantData, fileName, variables)
}
\arguments{
\item{relevantData}{FOO}

\item{fileName}{BAR}

\item{variables}{BAZ}
}
\description{
Something
}
\examples{
\dontrun{
# Generates REMIND emission report, extracts relevant variables and reshapes. Note: This funtion
can be used in a tidyverse pipeline.
emissionDataForClimateAssessment(
  remind2::reportEmi(fulldata.gdx),
  scenarioName = "SSP2EU-NPi-ar6",
  climateAssessmentYaml = file.path(
    system.file(package = "piamInterfaces"), "iiasaTemplates", "climate_assessment_variables.yaml"
  )
  logFile = "output/missing.log"
)
}
}
\author{
Tonn Rüter
}
