% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissionDataForClimateAssessment.R
\name{emissionDataForClimateAssessment}
\alias{emissionDataForClimateAssessment}
\title{emissionDataForClimateAssessment}
\usage{
emissionDataForClimateAssessment(
  remindEmissionReport,
  scenarioName,
  climateAssessmentYaml = NULL,
  logFile = "output/missing.log"
)
}
\arguments{
\item{remindEmissionReport}{Data frame containing REMIND emission data}

\item{scenarioName}{Name of the scenario}

\item{climateAssessmentYaml}{Path to the yaml file containing the variables needed for climate-assessment. If no
file path is given, the function gets the yaml file from the piamInterfaces package}

\item{logFile}{Path to the log file. Default is "output/missing.log"}
}
\value{
Data frame with the REMIND emission data reshaped for climate assessment
}
\description{
Converts remind emission data from long to wide format suitable for climate assessment. Only considers regions "GLO"
and "World" and extracts only the variables needed for climate assessment. Per default these are provided from the
AR6 mapping in the piamInterfaces package. The resulting data frame has one column for each year and one row for
each variable.
}
\examples{
\dontrun{
# Generates REMIND emission report, extracts relevant variables and reshapes. Note: This funtion
can be used in a tidyverse pipeline.
emissionDataForClimateAssessment(
  remind2::reportEmi(fulldata.gdx),
  scenarioName = "SSP2EU-NPi-ar6",
  climateAssessmentYaml = file.path(
    system.file(package = "piamInterfaces"), "iiasaTemplates", "climate_assessment_variables.yaml"
  )
  logFile = "output/missing.log"
)
}
}
\author{
Tonn Rüter
}
