% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissionImpulseSenarios.R
\name{emissionImpulseSenarios}
\alias{emissionImpulseSenarios}
\title{emissionImpulseSenarios}
\usage{
emissionImpulseSenarios(
  qf,
  var = "AR6 climate diagnostics|Infilled|Emissions|CO2|Energy and Industrial Processes",
  impulse = 3660,
  periods = c(2020, 2030, 2040, 2050, 2060, 2070, 2080, 2090, 2100, 2110, 2130),
  endYear = 2200,
  inbetween = 1
)
}
\arguments{
\item{qf}{\code{quitte} data frame containing the emission data from a single emission scenario}

\item{var}{Name of the variable to be perturbed}

\item{impulse}{Amount of additional emissions. Defaults to 3.66e2, CO2 equivalent in Mt to 1 GtC}

\item{periods}{Periods in which to add the emissions impulse. Defaults to c(2020, 2030, ..., 2110, 2130)}

\item{endYear}{Extends the original observation periods. Defaults to 2200}

\item{inbetween}{Duration of the extended observation periods in years. Defaults to 1}
}
\value{
\code{quitte} data frame containing all data for the emission impulse scenarios
}
\description{
Geenerates emission impulse scenarios from a harmonized and infilled emissions data by repeating the unperturbed
scenario as baseline and concatenating perturbed scenarios. Each scenario adds a specified quantity of additional
emissions to the spcified emission variable in the given time periods. The periods of the original data frame can be
extended to observe long term effects in the climate emulation
}
\examples{
\dontrun{
impulses <- read.quitte(cfg$harmInfEmissionsFile) \%>\%
  emissionImpulseSenarios() \%>\%
  write_csv(cfg$emissionsImpulseFile, quote = "none")
}
}
\author{
Tonn Rüter
}
