% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportConfFromYaml.R
\name{exportConfFromYaml}
\alias{exportConfFromYaml}
\title{exportConfFromYaml}
\usage{
exportConfFromYaml(yamlFileName = "default.yaml")
}
\arguments{
\item{yamlFileName}{Location of the configuration file. Default is 'which_var_in_which_file.yaml'}
}
\description{
Read yaml export configuration. It determines which variables to extract from the MAGICC7 climate assessment data &
to which file they are supposed to be written. Resulting data frame contains \code{magicc7Variable}, \code{gamsVariable} and
\code{fileName} columns.
}
\author{
Tonn Rüter
}
