% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMagiccVersion.R
\name{getMagiccVersion}
\alias{getMagiccVersion}
\title{getMagiccVersion}
\usage{
getMagiccVersion(magiccBin)
}
\arguments{
\item{magiccBin}{Path to the MAGICC binary to query}
}
\value{
Trimmed, single-line MAGICC version string
}
\description{
Obtains the MAGICC version by querying the supplied binary with \code{--version}.
Fails hard with a descriptive error if the version cannot be established.
}
\author{
Tonn Rüter
}
