% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remindClimateAssessment-package.R
\docType{package}
\name{remindClimateAssessment-package}
\alias{remindClimateAssessment}
\alias{remindClimateAssessment-package}
\title{remindClimateAssessment: REMIND integration of IIASA's `climate-assessment` package}
\description{
The REMIND integration of IIASA's `climate-assessment` provides a standardized interface to simple climate models such as MAGICC7.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/remindClimateAssessment}
}

}
\author{
\strong{Maintainer}: Tonn Rüter \email{tonn.rueter@pik-potsdam.de}

Authors:
\itemize{
  \item Tonn Rüter \email{tonn.rueter@pik-potsdam.de}
}

}
\keyword{internal}
