% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameVariableMagicc7ToRemind.R
\name{renameVariableMagicc7ToRemind}
\alias{renameVariableMagicc7ToRemind}
\title{renameVariableMagicc7ToRemind}
\usage{
renameVariableMagicc7ToRemind(varName)
}
\arguments{
\item{varName}{Variable to be renamed}
}
\value{
Renamed variable
}
\description{
Substitutes parts of MAGICC7 variable names to create REMIND compatible variable names
}
\author{
Tonn Rüter
}
