% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameVariableRemindToMagicc7.R
\name{renameVariableRemindToMagicc7}
\alias{renameVariableRemindToMagicc7}
\title{renameVariableRemindToMagicc7}
\usage{
renameVariableRemindToMagicc7(varName, magiccVersion = "7.5.3")
}
\arguments{
\item{varName}{Variable to be renamed}

\item{magiccVersion}{Version string of the MAGICC model in use. Default is "7.5.3"}
}
\value{
Renamed variable
}
\description{
Renames climate assessment variables given in the REMIND convetion MAGICC7 compatible variable names
}
\author{
Tonn Rüter
}
