% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTimeReport.R
\name{runTimeReport}
\alias{runTimeReport}
\title{Generate a report of run times.}
\usage{
runTimeReport(runTimes, prefix = "")
}
\arguments{
\item{runTimes}{A named vector of run times}

\item{prefix}{A string to prefix each line of the report. Default is "".}
}
\value{
A formatted string showing the duration of each run.
}
\description{
This function takes a named vector of run times and generates a formatted report showing the duration of each run.
The names of the run times must follow the pattern 'task_id_1 start' and 'task_id_1 end', where 'task_id_1'
can be any string identifier.
}
\examples{
\dontrun{
# Typical usage: Set up as empty vector
runTimes <- c()
# [... run set up ...]
runTimes <- c("task1 start" = Sys.time())
# [... run task1 ...]
runTimes <- c("task1 end" = Sys.time())
# [... something untimed in between ...]
runTimes <- c("task2 start" = Sys.time())
# [... run task2 ...]
runTimes <- c("task2 end" = Sys.time())
runTimeReport(runTimes)
}
}
