% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateAssessmentConfig.R
\name{climateAssessmentConfig}
\alias{climateAssessmentConfig}
\title{Configure Climate Assessment}
\usage{
climateAssessmentConfig(outputDir, mode)
}
\arguments{
\item{outputDir}{A REMIND output directory}

\item{mode}{Determines which probabiliy files is used. Must be either "report", "iteration", or a valid path}
}
\value{
A list containing the configuration settings for the climate assessment
}
\description{
Collects all necessary files and directories to set up climate assessment run from a REMIND output directory
}
\examples{
\dontrun{
climateAssessmentConfig("<outputDir>", "iteration") # yields
list(
  outputDir  = "/p/tmp/tonnru/remind/output/h_cpol_KLW_d50_2025-02-06_18.10.48",
  scenario   = "h_cpol_KLW_d50",
  condaEnv   = "/p/projects/rd3mod/python/environments/scm_magicc7_hpc",
  isArchived = FALSE,
  logFile    = "<outputDir>/log_climate.txt",
  workersDir = "<outputDir>/climate-assessment-data/workers",
  climateDir = "<outputDir>/climate-assessment-data",
  archiveDir = "",
  scriptsDir = "<Depending on your installation>",
  magiccBin  = "<Depending on your installation>",
  variablesFile = "<Depending on piamInterfaces>",
  infillingDatabase = "<Depending on your REMIND default.cfg>",
  probabilisticFile = "<Depending on your REMIND default.cfg>",
  nSets             = "<Depending on your REMIND probabilisticFile>",
  ...
)
}
}
